+++ doc_sphinx/notebooks_resources/ipython/profile_default/startup/00-pandas-options.py
@@ -1,0 +1,3 @@
+import pandas as pd
+
+pd.options.display.max_rows = 11