+++ doc_sphinx/static/use_cases/webapi_configuration.py
@@ -1,0 +1,7 @@
+# onetick-py WebAPI configuration for OneTick Cloud
+import os
+os.environ['OTP_WEBAPI'] = '1'
+os.environ['OTP_HTTP_ADDRESS'] = 'https://rest.cloud.onetick.com'
+os.environ['OTP_ACCESS_TOKEN_URL'] = 'https://cloud-auth.parent.onetick.com/realms/OMD/protocol/openid-connect/token'
+os.environ['OTP_CLIENT_ID'] = '__FILL_IN__'
+os.environ['OTP_CLIENT_SECRET'] = '__FILL_IN__'